removed isInternal and getChildren from Node interface - these methods can't be used within EJB (use accessor instead).

+++ de.uni_koeln.spinfo.strings.core/src/de/uni_koeln/spinfo/strings/algo/suffixtrees/node/memory/SuffixNode.java
@@ -107,7 +107,7 @@ public abstract class SuffixNode implements Node{
 	}
 
 
-	public int getTextNumber() {
+	public long getTextNumber() {
 		return textNumber;
 	}
